+++ src/main/java/repository/database/DatabaseConfig.java
@@ -1,4 +1,4 @@
-package database;
+package repository.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
