+++ src/main/java/repository/dto/DTO.java
@@ -1,4 +1,4 @@
-package service;
+package repository.dto;
 
 import java.util.Objects;
