+++ src/main/java/service/RoomService.java
@@ -1,4 +1,4 @@
-package services;
+package service;
 
 import exceptions.DuplicateRoomNameException;
 import exceptions.EmptyRoomNameException;
